control: clean up cache metadata after prune
